Check Terraformer::MultiLineString coordinates
Add some type checking of the coordinates parameter in
Terraformer::MultiLineString. Since a multilinestring must be an array
of an array of coordinates, we should be checking to make sure invalid
coordinates are not passed in.

Note that this change breaks things as there are some other issues that
need to be worked out with respect to the initialization of
MultiLineStrings.