Change how MultiLineStrings are created
As I'd found when writing tests for all the classes, MultiLineStrings
are initialized differently than the other classes. All other classes
can be initialized by the raw coordinates array from their result
geojson, but not MultiLineString. This also causes some issues when
trying to validate te structure of MultiLineStrings and the LineStrings
they create. Not sure exactly why this special logic was in
MultiLineString, but all other classes use Geometry::Primitive's
parsing. So if we just do that most things get fixed. Does change how
things are initialized, as is outlined in the changes to the
multi_line_string_spec.