Update how MultiLineStrings are compared
Given some of the other changes, a test error appeared that had to do
with sorting a MultiLineString's constituent line_strings. It boiled
down to the fact that LineString doesn't have <=> defined, which is what
is being used by sort. However, we don't really need to add this, as the
class that does have <=> defined is Coordinate. Really we are just
sorting LineStrings based roughly on their first coordinate.

Honestly I don't really know how or why this code was supposed to be
working previously.