Pull iter_coordinates into Geometry
The coordinates iteration methods shouldn't be on Enumerable. Refactored
this code to live as a class method on Geometry.

Geometry.iter_coordinate accepts the object to be iterated on, the
method and a block. If it sees a Coordinate or a non-Array object it
will yield that object to the block. Otherwise it recursively iterates
on the Array.

Geometry now has instance methods each_coordinate and map_coordinate
that run iter_coordinate on the coordinates object of the Geometry.

The only kinda ugly part is that ConvexHull now has to call
Geometry.iter_coordinates explicitly. But this is a small price for not
polluting the Enumerable namespace anymore.